rename: RouterMap 이름 변경
앞에 공백이 있던 문제를 해결하기 위해 공백을 제거하였습니다
